requirements.txt
# FastAPI y servidor
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Base de datos PostgreSQL 17 (async)
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0

# Configuración y validación
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0

# Migraciones (opcional)
alembic>=1.14.0
psycopg2-binary>=2.9.10

# Utilidades
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.10.0
